--- src/http-client/types.ts
@@ -1,8 +1,0 @@
-import type { AxiosRequestConfig } from 'axios';
-import type { AxiosInstanceWrapper } from 'middleware-axios';
-
-export type HttpClient = AxiosInstanceWrapper;
-
-export interface HttpClientFactory {
-  (config?: AxiosRequestConfig): HttpClient;
-}
