+++ src/preset/node/node/http-client/__test__/index.test.ts
@@ -1,7 +1,70 @@
 import { Context } from '@opentelemetry/api';
 import { SemanticAttributes } from '@opentelemetry/semantic-conventions';
 import { Span, Tracer } from '@opentelemetry/sdk-trace-base';
-import { tracingMiddleware, getRequestInfo, hideFirstId } from '../tracing';
+import { tracingMiddleware, getRequestInfo, hideFirstId, displayUrl } from '..';
+
+describe('displayUrl', () => {
+  it('should properly merge baseURL with url', () => {
+    const cases: Array<{ url: string; baseURL: string; expectedUrl: string }> = [
+      // baseURL и url
+      {
+        baseURL: 'https://www.base.com/',
+        url: '/user/current',
+        expectedUrl: 'https://www.base.com/user/current',
+      },
+      {
+        baseURL: 'https://www.base.com',
+        url: 'user/current',
+        expectedUrl: 'https://www.base.com/user/current',
+      },
+      {
+        baseURL: 'https://www.base.com',
+        url: '/user/current',
+        expectedUrl: 'https://www.base.com/user/current',
+      },
+      {
+        baseURL: 'https://www.base.com/',
+        url: 'admin/all',
+        expectedUrl: 'https://www.base.com/admin/all',
+      },
+
+      // только baseURL
+      {
+        baseURL: 'www.test.com',
+        url: '',
+        expectedUrl: 'www.test.com',
+      },
+      {
+        baseURL: 'www.test.com/',
+        url: '',
+        expectedUrl: 'www.test.com/',
+      },
+
+      // только url
+      {
+        baseURL: '',
+        url: '/hello/world',
+        expectedUrl: '/hello/world',
+      },
+      {
+        baseURL: '',
+        url: 'some/path',
+        expectedUrl: 'some/path',
+      },
+
+      // ничего
+      {
+        baseURL: '',
+        url: '',
+        expectedUrl: '[empty]',
+      },
+    ];
+
+    for (const { baseURL, url, expectedUrl } of cases) {
+      expect(displayUrl(baseURL, url)).toBe(expectedUrl);
+    }
+  });
+});
 
 describe('tracingMiddleware', () => {
   it('should handle success response', async () => {
