+++ src/tokens.ts
@@ -5,7 +5,6 @@ import type * as express from 'express';
 import type { ConfigSource, BaseConfig } from './config/types';
 import type { Logger } from './log/types';
 import type { Cache } from './cache/types';
-import type { HttpClientFactory } from './http-client/types';
 import type { LogMiddlewareHandlerInit } from './http-client/middleware/log';
 import type { SagaExtendedMiddleware } from './utils/redux-saga';
 import type { BridgeClientSide, BridgeServerSide } from './utils/ssr';
@@ -13,10 +12,10 @@ import type { Tracer } from '@opentelemetry/api';
 import type { BasicTracerProvider, SpanExporter } from '@opentelemetry/sdk-trace-base';
 import type { Resource } from '@opentelemetry/resources';
 import type { ElementType, ReactNode } from 'react';
-
-// @todo придумать как брать не из пресетов:
 import type { PageAssets, HandlerContext, StrictMap, KnownHttpApiKey } from './preset/parts/types';
 import type { SpecificExtras } from './preset/node/handler';
+import type { CreateAxiosDefaults } from 'axios';
+import type { AxiosInstanceWrapper } from 'middleware-axios';
 
 export const KnownToken = {
   // config
@@ -58,7 +57,8 @@ export const KnownToken = {
 
     // @todo переименовать в Axios?
     Client: {
-      factory: createToken<HttpClientFactory>('client/factory'),
+      factory:
+        createToken<(defaults?: CreateAxiosDefaults) => AxiosInstanceWrapper>('client/factory'),
       Middleware: {
         Log: {
           handler: createToken<LogMiddlewareHandlerInit>('log/handler'),
